+++ src/app/budget/budget-dashboard/budget-dashboard.component.ts
@@ -1,4 +1,5 @@
 import { Component, OnInit, Input, HostListener, Inject } from '@angular/core';
+import { Router } from '@angular/router';
 import { TranslateService } from '@ngx-translate/core';
 import { Observable } from 'rxjs/Observable';
 import { Budget } from '../../models/budget.model';
@@ -12,6 +13,8 @@ import * as UiStateActions from '../../store/actions/uiState.actions';
 import * as fromRoot from '../../store/app.reducers';
 import * as fromBudgetApp from '../store/reducers/index';
 import { DOCUMENT } from '@angular/platform-browser';
+import { MatDialog } from '@angular/material';
+import { BudgetLinesComponent } from '../budget-lines/budget-lines.component';
 
 @Component({
   selector: 'app-budget-dashboard',
@@ -44,7 +47,8 @@ export class BudgetDashboardComponent implements OnInit, OnDestroy {
   userSubscription: Subscription;
 
   constructor(translate: TranslateService,
-    private store: Store<fromRoot.AppState>) {
+    private store: Store<fromRoot.AppState>,
+    private router: Router) {
     this.user$ = this.store.select(fromRoot.selectUser);
     this.budget$ = this.store.select(fromBudgetApp.selectCurrentBudget);
 
@@ -55,15 +59,17 @@ export class BudgetDashboardComponent implements OnInit, OnDestroy {
     this.store.dispatch(new UiStateActions.ChangeTitleAction('budgettitle'));
     this.store.dispatch(new UiStateActions.ChangeMainMenuBtnVisibleAction(true));
     this.lastOffset = 0;
-
-
-
   }
 
   ngOnDestroy() {
     if (this.userSubscription) {
       this.userSubscription.unsubscribe();
     }
+  }
+
+  onAddFastExpense(): void {
+    this.router.navigate(['addFastExpense']);
+
   }