+++ src/app/budget/budget.component.ts
@@ -41,6 +41,7 @@ export class BudgetComponent implements OnInit, OnDestroy {
 
     this.store.dispatch(new UiStateActions.ChangeTitleAction('budgettitle'));
     this.store.dispatch(new UiStateActions.ChangeMainMenuBtnVisibleAction(true));
+    this.store.dispatch(new UiStateActions.ChangeMainToolbarCloseBtnVisibleAction(false));
 
     this.userSubscription = this.user$.subscribe(user => {
       // console.log('userSubscription: ', user);