+++ src/app/budget/store/reducers/index.spec.ts
@@ -1,0 +1,36 @@
+import * as fromModuleReducer from './index';
+import * as fromBudget from './budget.reducer';
+import * as fromBudgetLine from './budget-line.reducer';
+import * as fromExpense from './expense.reducer';
+
+describe('BudgetModule selectors', () => {
+    const initialState: fromModuleReducer.State = {
+        budgets: fromBudget.initialState,
+        budgetLines: fromBudgetLine.initialState,
+        expenses: fromExpense.initialState
+    }
+    describe('feature selectors', () => {
+        it('getBudgetModuleState should be initialized', () => {
+            expect(fromModuleReducer.getBudgetModuleState).toBeDefined();
+        });
+        it('getBudgetsState should be initialized', () => {
+            expect(fromModuleReducer.getBudgetsState).toBeDefined();
+        });
+        it('getBudgetLinesState should be initialized', () => {
+            expect(fromModuleReducer.getBudgetLinesState).toBeDefined();
+        });
+        it('getExpensesState should be initialized', () => {
+            expect(fromModuleReducer.getExpensesState(initialState)).toBeDefined();
+        });
+
+    });
+
+
+
+    describe('state selectors', () => {
+
+    });
+
+});
+
+