+++ src/app/core/components/toolbar/toolbar.component.ts
@@ -4,6 +4,7 @@ import { Store } from '@ngrx/store';
 import * as fromRoot from '../../../store/app.reducers';
 import * as UiStateActions from '../../../store/actions/uiState.actions';
 import { OnInit } from '@angular/core';
+import { Location } from '@angular/common';
 
 @Component({
   selector: 'app-toolbar',
@@ -13,14 +14,21 @@ import { OnInit } from '@angular/core';
 export class ToolbarComponent implements OnInit {
   @Output() switchSideMenu = new EventEmitter();
   uiTitle$: Observable<string>;
+  closeBtnVisible$: Observable<boolean>;
 
 
-  constructor(private store: Store<fromRoot.AppState>) {
+  constructor(private store: Store<fromRoot.AppState>,
+    private _location: Location) {
     this.uiTitle$ = this.store.select(fromRoot.selectUiTitle);
+    this.closeBtnVisible$ = this.store.select(fromRoot.selectUiMainToolbarCloseBtnVisible);
   }
 
   ngOnInit() {
 
+  }
+
+  onCloseBtn() {
+    this._location.back();
   }